chore: index.tsx 제거 및 파일명을 컴포넌트명.tsx로 변경
