+++ src/chat/components/Chat.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { ChatMessagesProvider } from "@/chat/hooks/useChatMessagesStore";
-import ChatHeader from "../ChatHeader";
-import ChatRoom from "../ChatRoom";
+import ChatHeader from "./ChatHeader";
+import ChatRoom from "./ChatRoom";
 
 export default function Chat() {
   // TODO: 채팅 메세지 목록 프리페치 SSR 필요
