+++ src/chat/components/ChatAvatar.tsx
