+++ src/chat/components/ChatBubble.tsx
