+++ src/chat/components/TextareaAutoSize.tsx
