+++ src/shared/components/MainContent.tsx
