+++ src/shared/components/ResponsiveRootLayout.tsx
@@ -3,11 +3,7 @@
 import { ReactNode } from "react";
 import { css } from "styled-components";
 
-export default function ResponsiveRootLayout({
-  children,
-}: {
-  children: ReactNode;
-}) {
+export default function ResponsiveRootLayout({ children }: { children: ReactNode }) {
   return (
     <div
       css={css`
